refactor: rename folder lowcase
