+++ front-end/mobile/lib/pages/welcomePage/reg_screen.dart
