+++ front-end/mobile/lib/pages/welcomePage/typewriter_text.dart
