+++ front-end/mobile/lib/pages/welcomePage/welcome_screen.dart
